Move project machine to fsm (finite-state machine) folder
