+++ app/assets/scripts/fsm/project/actions.js
