+++ app/assets/scripts/fsm/project/helpers.js
