+++ app/assets/scripts/fsm/project/index.js
