+++ app/assets/scripts/fsm/project/machine.js
