+++ app/assets/scripts/fsm/project/websocket-client.js
